Entrega Final Hay muchos problemas de seguridad que aun hay que arreglar

+++ cybercloud-backend/src/controllers/itempurchased.controller.js
@@ -20,28 +20,35 @@ itempurchasedCtrl.getItemPurchased = async(req,res)=> {
     }
 }
 itempurchasedCtrl.createItemPurchased = async (req,res)=> {
-    try {
-        const {items, user, username, operator, operatorname, id} = req.body;
-        //console.log(req.body)
-        for (let index = 0; index < items.length; index++) {
-            console.log(items[index])
-            const newItemP = new ItemPurchased({
-                id_item: items[index].article,
-                description: items[index].description,
-                price: items[index].price,
-                amount: items[index].amount,
-                numberBox: global.setting.numberBox,
-                user,
-                username,
-                operator,
-                operatorname
-            })
-            await newItemP.save();
+    if (global.setting.boxStatus) {
+        try {
+            const {items, user, username, operator, operatorname, id} = req.body;
+            //console.log(req.body)
+            for (let index = 0; index < items.length; index++) {
+                console.log(items[index])
+                const newItemP = new ItemPurchased({
+                    id_item: items[index].article,
+                    description: items[index].description,
+                    price: items[index].price,
+                    amount: items[index].amount,
+                    numberBox: global.setting.numberBox,
+                    user,
+                    username,
+                    operator,
+                    operatorname
+                })
+                await newItemP.save();
+            }
+            await Petition.findByIdAndDelete(id);
+            return res.status(200).send({status:true})
+        } catch (error) {
+            return res.status(400).send({status:false, error})
         }
-        await Petition.findByIdAndDelete(id);
-        return res.status(200).send({status:true})
-    } catch (error) {
-        return res.status(400).send({status:false, error})
+    }
+    else{
+        return res.send({
+            error: "Box is Closed."
+        });
     }
 }
 itempurchasedCtrl.deleteItemPurchased = async (req,res)=> {
